move stages to video subfolder
