+++ ray-curator/ray_curator/stages/video/clipping/clip_extraction_stages.py
