+++ ray-curator/ray_curator/stages/video/clipping/frame_extraction.py
